Rename media field internal file to component.tsx for consistency.
