+++ src/packages/admin-ui-components/src/detail-panel/fields/media-field/component.tsx
