+++ src/packages/admin-ui-components/src/detail-panel/fields/media-field/index.ts
@@ -1,1 +1,1 @@
-export * from './media-field';
+export * from './component';
